Clean up `require.cache` handling
Follow-up from #68535.

Since the `require.cache` handling is not specific for Webpack, we move
the helper functions out of `nextjs-require-cache-hot-reloader.ts`.

In addition, when deleting an entry from the `require.cache`, while
cleaning up its references in the `children` of parent modules, we now
consider all entries in the `require.cache` instead of relying on a list
of "origin modules". This list had entries that were not needed, most
notably the server runtime bundles (the experimental runtimes were even
missing), and at least one (`node-manifest-loader`) was also missing.

This has a negligible effect on performance. Handling 1200
`require.cache` entries takes 0.3ms on an M2. The approach was also
previously aligned with @sokra.