refactor(turbo-tasks): Add TaskPersistence enum for task creation functions (#68866)
Builds on top of the `is_transient` boolean arguments from #68436, but
uses an enum to provide a third variant of
`TaskPersistence::LocalCells`.

The branches for this are `todo!()` right now, but the idea is that
we'll intercept these in the manager and schedule them inside the
manager instead of creating a unique task with its own `TaskId` on the
backend.

## Testing

Some of the `local_cell` tests are broken with this change, as functions
annotated with `#[turbo_tasks::function(local_cells)]` now hit the
`todo!()` blocks added in this PR. Those tests are disabled for now, and
I'll fix this in a later PR. *(This is okay, nothing is using local
cells yet, aside from those tests)*