[devlow-bench] fail with non-zero exit and restore file.js
devlow-bench regressed, and it went unnoticed since the cli simply logged a stack and exited the process normally, so CI continued passing. This corrects that by exiting the process with a non-zero code.

This also fixes the cause of the regression, where `file.js` was not brought over during the repo merge.

Test Plan: devlow-bench job in CI.

[0] https://github.com/vercel/next.js/blob/cc484b996ad235c4a5acc284c099c2d5cab8c72d/turbopack/packages/devlow-bench/src/cli.ts#L112-L114