bugfix: ensure prefetch aliasing doesn't apply URL redirect (#69153)
Another follow up to #68340: When a prefetch is aliased (meaning we
returned a prefetch entry that corresponded with the same pathname, but
different / missing search params), we signal to the router that it
should only copy the loading state from this prefetch entry, and null
everything else out so it can be fetched from the server. However it's
possible that we return an aliased entry that corresponds with a
redirect (such as in middleware) meaning the server will supply a
`canonicalUrlOverride` value that will correspond with the aliased
entry, rather than the requested entry.

This updates the handling to only apply the canonical URL override when
we're not using an aliased entry. This will defer the redirect handling
to when the actual dynamic request (in the server patch reducer) so that
the aliased entry will only handle the loading copying and not do any
redirects.

This copies the existing tests that covered this behavior but adds a
middleware redirect to verify the behavior is the same.