Ensure fetch cache TTL is updated properly (#69164)
To reduce the number of set requests we were sending upstream we added
an optimization to skip sending the set request if the cached data
didn't change after a revalidation. The problem with this optimization
is the upstream service relies on this set request to know to update the
cache entries TTL and consider it up to date. This causes unexpected
revalidations while the cache value hasn't changed and the TTL is kept
stale.

x-ref: NEXT-3693