Set max retries for build failures to zero
We generally a failure in a build is a sign of an actual problem. JestWorker has a default retries of 3 which means for every failing error we atempt it 4 times before giving up. This update changes the max retires to 0 so that we abandon the build after the first failure.