Remove obsolete `.react-version` file
This file was only added to automatically add GH labels.
However, we now set this label in the automated GH workflow.
It was also problematic since the labeler only ever adds labels.
Sometimes the label was only added during merge or intermediate steps but not on the final PR making the label not useful as a single signal.