refactor path creation in walkTreeWithFlightRouterState (#69383)
There's no reason to use `map` here, since it's making the code harder
to read and we ultimately have to flatten & filter out falsey values.

This updates the path construction to use a `for..of` and only pushes
paths that are relevant. This also removes the nested path array that
was getting created every iteration of the loop.