chore(CI): Use Graphite's CI Optimization Feature (#69204)
## What?

https://graphite.dev/blog/ci-optimizer
https://graphite.dev/docs/stacking-and-ci

> Stacking leads to developers creating smaller easier-to-review pull requests, which can lead to more CI runs unless you optimize your CI for stacking. [...]
>
> Additional CI runs occur when stacking PRs in part due to to the additional PRs created, and due to behind-the-scenes rebasing to keep stacked branches up to date.
>
> To solve this, Graphite offers first-class integrations with [...] GitHub Actions to let you customize which PRs in stacks you want to run CI on.

I've configured CI optimization to run on the top & bottom commits of each stack:

![Screenshot 2024-08-22 at 3.33.35 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/HAZVitxRNnZz8QMiPn4a/67c9e2a6-92f4-4726-bb08-3b9df246ed9b.png)

## Failure Handling

If this fails for some reason, the full CI suite should run.

> Graphite's [...] GitHub Actions integrations are configured to "fail open" so that outages and errors still result in your CI running.

## Opting Out

Add the new `CI Bypass Graphite Optimization` label to your PR, and then push an update to the branch.

**Ugly:** You must push an update to the branch. Triggering on label change would cause unintended side-effects, and just triggering the workflow to re-run in the GitHub UI doesn't cause the workflow to pick up the changed label, as the label is part of the pull request event.

## What about `build_and_deploy`?

Because this currently triggers off of a `push` event rather than a `pull_request` (allowing preview deployments without making a PR), there's currently no way to change the behavior on `build_and_deploy`.

This is implemented as a reusable workflow, so it should be possible to hook this into other workflows in the future (e.g. if we wanted to avoid running `pull_request_stats` as well), as long as they trigger on `pull_request` events.

## Testing

### Breakage on Top of Stack

Push a PR to the top of the stack that breaks an e2e test. See that CI fails on that PR as expected, and that all of the appropriate CI jobs run.

![Screenshot 2024-08-26 at 16-08-14 please-ignore(CI) Break e2e test (tests #69204) · vercel_next js@fb33205](https://github.com/user-attachments/assets/f62698df-1dcc-45bf-ac1c-7cf2ca17af83)

### Breakage in Middle of Stack

Push another PR on top of the stack, so that the e2e breakage is in the middle of the stack. Force github to re-run the workflow using the github UI. See that the CI optimizer skips the PR because it's in the middle:

<img width="1024" alt="Screenshot 2024-08-26 at 4 17 52 PM" src="https://github.com/user-attachments/assets/3fd2c473-64e0-4c7c-8169-438fbb60e758">

And see that most jobs are skipped, leaving us with a very small set of test jobs:

<img width="875" alt="Screenshot 2024-08-26 at 4 21 53 PM" src="https://github.com/user-attachments/assets/91623d84-b4b7-4d00-8177-37ffbf773046">


### Observed Failure at Top of Stack

Even though the failure is caused by the middle of the stack, a documentation-only change at the top of the stack still catches the failure caused by the middle of the stack, as the "documentation only" check compares against canary, not just the parent PR in the stack.

<img width="1713" alt="Screenshot 2024-08-26 at 4 24 27 PM" src="https://github.com/user-attachments/assets/dd2ba177-8579-42c7-a602-56d03732f44d">


### Bypassing CI Optimization

Add the new label to the PR in the middle of the stack, and force it to re-run by making a change to the commit message and re-submitting with `gh submit`.

<img width="468" alt="Screenshot 2024-08-26 at 4 26 49 PM" src="https://github.com/user-attachments/assets/c04b0379-b4f4-4d83-b41b-a1aa3a297f8d">

See that the debug output shows the label being detected:

<img width="416" alt="Screenshot 2024-08-26 at 4 35 11 PM" src="https://github.com/user-attachments/assets/a2828f49-7a76-4266-ba7c-de370045bcb5">

And see that the full CI suite runs!