bugfix: aliased prefetches should flow through PPR handling (#69476)
I was previously incorrectly excluding aliased entries from flowing
through the PPR path. However, since accessing `searchParams` will
postpone, the aliased entry is safe to flow through the PPR branch. When
PPR is enabled and the non-PPR branch attempts to handle this
navigation, it won't know how to handle the postponed response.

Separately, I noticed that the PPR branch was not spawning a pending
task if the page segment contained search params, because it was
checking for strict equality, when in reality it'd be something like
`__PAGE__?{'foo':'bar'}`

Fixes #69437
Closes NDX-255